pyproject.toml
[tool.poetry]
name = "artefacts"
version = "2.0.0"
description = "Python utilities for interacting with dbt artifacts"
authors = ["Tom Waterman <tjwaterman99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.2"
dbt-core = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
dbt-postgres = "^1.1.0"
ipython = "7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
